+++ frame/election-provider-support/src/lib.rs
@@ -100,6 +100,7 @@
 //!     pub struct Module<T: Config>(std::marker::PhantomData<T>);
 //!
 //!     impl<T: Config> ElectionDataProvider<AccountId, BlockNumber> for Module<T> {
+//!         const MAXIMUM_VOTES_PER_VOTER: u32 = 1;
 //!         fn desired_targets() -> data_provider::Result<(u32, Weight)> {
 //!             Ok((1, 0))
 //!         }
@@ -180,6 +181,9 @@ pub mod data_provider {
 
 /// Something that can provide the data to an [`ElectionProvider`].
 pub trait ElectionDataProvider<AccountId, BlockNumber> {
+	/// Maximum number of votes per voter that this data provider is providing.
+	const MAXIMUM_VOTES_PER_VOTER: u32;
+
 	/// All possible targets for the election, i.e. the candidates.
 	///
 	/// If `maybe_max_len` is `Some(v)` then the resulting vector MUST NOT be longer than `v` items
@@ -226,6 +230,7 @@ pub trait ElectionDataProvider<AccountId, BlockNumber> {
 
 #[cfg(feature = "std")]
 impl<AccountId, BlockNumber> ElectionDataProvider<AccountId, BlockNumber> for () {
+	const MAXIMUM_VOTES_PER_VOTER: u32 = 0;
 	fn targets(_maybe_max_len: Option<usize>) -> data_provider::Result<(Vec<AccountId>, Weight)> {
 		Ok(Default::default())
 	}